Introduce non-blocking mode for the command's output in the C program
So that the user can continue to navigate through the output of the previous executing while the current one is executing. This improves experience when commands take a long time to execute.

Also change the title variables such that the titles reflect the state at the beginning of command executing (especially useful to know the time of the last refresh).